Cell Wall Gene Evo2 Analysis/pyproject.toml
[project]
name = "cell-wall-gene-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "biopython>=1.86",
    "matplotlib>=3.10.8",
    "polars>=1.39.3",
    "scikit-learn>=1.8.0",
]
